+++ erizo/src/test/rtp/RtpRetransmissionHandlerTest.cpp
@@ -10,11 +10,9 @@
 #include <string>
 #include <vector>
 
-static constexpr uint16_t kVideoSsrc = 1;
-static constexpr uint16_t kAudioSsrc = 2;
-static constexpr uint16_t kArbitrarySeqNumber = 12;
-static constexpr uint16_t kFirstSequenceNumber = 0;
-static constexpr uint16_t kLastSequenceNumber = 65535;
+#include "../utils/Mocks.h"
+#include "../utils/Tools.h"
+#include "../utils/Matchers.h"
 
 using ::testing::_;
 using ::testing::IsNull;
@@ -33,32 +31,6 @@ using erizo::InboundHandler;
 using erizo::OutboundHandler;
 using erizo::Worker;
 
-class MockWebRtcConnection: public WebRtcConnection {
- public:
-  MockWebRtcConnection(std::shared_ptr<Worker> worker, const IceConfig &ice_config,
-                       const std::vector<RtpMap> rtp_mappings) :
-    WebRtcConnection(worker, "", ice_config, rtp_mappings, nullptr) {}
-
-  virtual ~MockWebRtcConnection() {
-  }
-};
-
-class Reader : public InboundHandler {
- public:
-  MOCK_METHOD0(enable, void());
-  MOCK_METHOD0(disable, void());
-  MOCK_METHOD0(getName, std::string());
-  MOCK_METHOD2(read, void(Context*, std::shared_ptr<dataPacket>));
-};
-
-class Writer : public OutboundHandler {
- public:
-  MOCK_METHOD0(enable, void());
-  MOCK_METHOD0(disable, void());
-  MOCK_METHOD0(getName, std::string());
-  MOCK_METHOD2(write, void(Context*, std::shared_ptr<dataPacket>));
-};
-
 class RtpRetransmissionHandlerTest : public ::testing::Test {
  public:
   RtpRetransmissionHandlerTest() : ice_config() {}
@@ -67,15 +39,15 @@ class RtpRetransmissionHandlerTest : public ::testing::Test {
   virtual void SetUp() {
     scheduler = std::make_shared<Scheduler>(1);
     worker = std::make_shared<Worker>(scheduler);
-    connection = std::make_shared<MockWebRtcConnection>(worker, ice_config, rtp_maps);
+    connection = std::make_shared<erizo::MockWebRtcConnection>(worker, ice_config, rtp_maps);
 
-    connection->setVideoSinkSSRC(kVideoSsrc);
-    connection->setAudioSinkSSRC(kAudioSsrc);
+    connection->setVideoSinkSSRC(erizo::kVideoSsrc);
+    connection->setAudioSinkSSRC(erizo::kAudioSsrc);
 
     pipeline = Pipeline::create();
     rtx_handler = std::make_shared<RtpRetransmissionHandler>(connection.get());
-    reader = std::make_shared<Reader>();
-    writer = std::make_shared<Writer>();
+    reader = std::make_shared<erizo::Reader>();
+    writer = std::make_shared<erizo::Writer>();
 
     pipeline->addBack(writer);
     pipeline->addBack(rtx_handler);
@@ -86,69 +58,38 @@ class RtpRetransmissionHandlerTest : public ::testing::Test {
   virtual void TearDown() {
   }
 
-  std::shared_ptr<dataPacket> createDataPacket(uint16_t seq_number, packetType type) {
-    erizo::RtpHeader *header = new erizo::RtpHeader();
-    header->setSeqNumber(seq_number);
-
-    if (type == AUDIO_PACKET) {
-      header->setSSRC(kAudioSsrc);
-    } else {
-      header->setSSRC(kVideoSsrc);
-    }
-
-    return std::make_shared<dataPacket>(0, reinterpret_cast<char*>(header), sizeof(erizo::RtpHeader), type);
-  }
-
-  std::shared_ptr<dataPacket> createNack(uint16_t seq_number, packetType type, int additional_packets = 0) {
-    erizo::RtcpHeader *nack = new erizo::RtcpHeader();
-    nack->setPacketType(RTCP_RTP_Feedback_PT);
-    nack->setBlockCount(1);
-    uint source_ssrc = type   == VIDEO_PACKET ? connection->getVideoSinkSSRC() : connection->getAudioSinkSSRC();
-    uint ssrc = type == VIDEO_PACKET ? connection->getVideoSourceSSRC() : connection->getAudioSourceSSRC();
-    nack->setSSRC(ssrc);
-    nack->setSourceSSRC(source_ssrc);
-    nack->setNackPid(seq_number);
-    nack->setNackBlp(additional_packets);
-    nack->setLength(3);
-    char *buf = reinterpret_cast<char*>(nack);
-    int len = (nack->getLength() + 1) * 4;
-    return std::make_shared<dataPacket>(0, buf, len, type);
-  }
-
   IceConfig ice_config;
   std::vector<RtpMap> rtp_maps;
-  std::shared_ptr<MockWebRtcConnection> connection;
+  std::shared_ptr<erizo::MockWebRtcConnection> connection;
   Pipeline::Ptr pipeline;
-  std::shared_ptr<Reader> reader;
-  std::shared_ptr<Writer> writer;
+  std::shared_ptr<erizo::Reader> reader;
+  std::shared_ptr<erizo::Writer> writer;
   std::shared_ptr<RtpRetransmissionHandler> rtx_handler;
   std::shared_ptr<Worker> worker;
   std::shared_ptr<Scheduler> scheduler;
 };
 
-MATCHER_P(HasSequenceNumber, seq_num, "") {
-  return (reinterpret_cast<erizo::RtpHeader*>(std::get<0>(arg)->data))->getSeqNumber() == seq_num;
-}
-
 TEST_F(RtpRetransmissionHandlerTest, basicBehaviourShouldReadPackets) {
-    auto packet = createDataPacket(kArbitrarySeqNumber, VIDEO_PACKET);
+    auto packet = erizo::Tools::createDataPacket(erizo::kArbitrarySeqNumber, VIDEO_PACKET);
 
-    EXPECT_CALL(*reader.get(), read(_, _)).With(Args<1>(HasSequenceNumber(kArbitrarySeqNumber))).Times(1);
+    EXPECT_CALL(*reader.get(), read(_, _)).With(Args<1>(HasSequenceNumber(erizo::kArbitrarySeqNumber))).Times(1);
     pipeline->read(packet);
 }
 
 TEST_F(RtpRetransmissionHandlerTest, basicBehaviourShouldWritePackets) {
-    auto packet = createDataPacket(kArbitrarySeqNumber, VIDEO_PACKET);
+    auto packet = erizo::Tools::createDataPacket(erizo::kArbitrarySeqNumber, VIDEO_PACKET);
 
-    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(kArbitrarySeqNumber))).Times(1);
+    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(erizo::kArbitrarySeqNumber))).Times(1);
     pipeline->write(packet);
 }
 
 TEST_F(RtpRetransmissionHandlerTest, shouldRetransmitPackets_whenReceivingNacksWithGoodSeqNum) {
-    auto rtp_packet = createDataPacket(kArbitrarySeqNumber, VIDEO_PACKET);
-    auto nack_packet = createNack(kArbitrarySeqNumber, VIDEO_PACKET);
+    auto rtp_packet = erizo::Tools::createDataPacket(erizo::kArbitrarySeqNumber, VIDEO_PACKET);
+    uint ssrc = connection->getVideoSourceSSRC();
+    uint source_ssrc = connection->getVideoSinkSSRC();
+    auto nack_packet = erizo::Tools::createNack(ssrc, source_ssrc, erizo::kArbitrarySeqNumber, VIDEO_PACKET);
 
-    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(kArbitrarySeqNumber))).Times(2);
+    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(erizo::kArbitrarySeqNumber))).Times(2);
     pipeline->write(rtp_packet);
 
     EXPECT_CALL(*reader.get(), read(_, _)).Times(0);
@@ -156,11 +97,13 @@ TEST_F(RtpRetransmissionHandlerTest, shouldRetransmitPackets_whenReceivingNacksW
 }
 
 TEST_F(RtpRetransmissionHandlerTest, shouldNotRetransmitPackets_whenReceivingNacksWithBadSeqNum) {
-    auto rtp_packet = createDataPacket(kArbitrarySeqNumber, VIDEO_PACKET);
-    auto nack_packet = createNack(kArbitrarySeqNumber + 1, VIDEO_PACKET);
+    auto rtp_packet = erizo::Tools::createDataPacket(erizo::kArbitrarySeqNumber, VIDEO_PACKET);
+    uint ssrc = connection->getVideoSourceSSRC();
+    uint source_ssrc = connection->getVideoSinkSSRC();
+    auto nack_packet = erizo::Tools::createNack(ssrc, source_ssrc, erizo::kArbitrarySeqNumber + 1, VIDEO_PACKET);
 
-    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(kArbitrarySeqNumber))).Times(1);
-    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(kArbitrarySeqNumber + 1))).Times(0);
+    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(erizo::kArbitrarySeqNumber))).Times(1);
+    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(erizo::kArbitrarySeqNumber + 1))).Times(0);
     pipeline->write(rtp_packet);
 
     EXPECT_CALL(*reader.get(), read(_, _)).Times(1);
@@ -168,10 +111,12 @@ TEST_F(RtpRetransmissionHandlerTest, shouldNotRetransmitPackets_whenReceivingNac
 }
 
 TEST_F(RtpRetransmissionHandlerTest, shouldNotRetransmitPackets_whenReceivingNacksFromDifferentType) {
-    auto rtp_packet = createDataPacket(kArbitrarySeqNumber, VIDEO_PACKET);
-    auto nack_packet = createNack(kArbitrarySeqNumber, AUDIO_PACKET);
+    auto rtp_packet = erizo::Tools::createDataPacket(erizo::kArbitrarySeqNumber, VIDEO_PACKET);
+    uint ssrc = connection->getAudioSourceSSRC();
+    uint source_ssrc = connection->getAudioSinkSSRC();
+    auto nack_packet = erizo::Tools::createNack(ssrc, source_ssrc, erizo::kArbitrarySeqNumber, AUDIO_PACKET);
 
-    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(kArbitrarySeqNumber))).Times(1);
+    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(erizo::kArbitrarySeqNumber))).Times(1);
     pipeline->write(rtp_packet);
 
     EXPECT_CALL(*reader.get(), read(_, _)).Times(1);
@@ -179,36 +124,44 @@ TEST_F(RtpRetransmissionHandlerTest, shouldNotRetransmitPackets_whenReceivingNac
 }
 
 TEST_F(RtpRetransmissionHandlerTest, shouldRetransmitPackets_whenReceivingWithSeqNumBeforeGeneralRollover) {
-    auto nack_packet = createNack(kFirstSequenceNumber, VIDEO_PACKET);
+    uint ssrc = connection->getVideoSourceSSRC();
+    uint source_ssrc = connection->getVideoSinkSSRC();
+    auto nack_packet = erizo::Tools::createNack(ssrc, source_ssrc, erizo::kFirstSequenceNumber, VIDEO_PACKET);
 
-    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(kLastSequenceNumber))).Times(1);
-    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(kFirstSequenceNumber))).Times(2);
-    pipeline->write(createDataPacket(kLastSequenceNumber, VIDEO_PACKET));
-    pipeline->write(createDataPacket(kFirstSequenceNumber, VIDEO_PACKET));
+    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(erizo::kLastSequenceNumber))).Times(1);
+    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(erizo::kFirstSequenceNumber))).Times(2);
+    pipeline->write(erizo::Tools::createDataPacket(erizo::kLastSequenceNumber, VIDEO_PACKET));
+    pipeline->write(erizo::Tools::createDataPacket(erizo::kFirstSequenceNumber, VIDEO_PACKET));
 
     EXPECT_CALL(*reader.get(), read(_, _)).Times(0);
     pipeline->read(nack_packet);
 }
 
 TEST_F(RtpRetransmissionHandlerTest, shouldRetransmitPackets_whenReceivingWithSeqNumBeforeBufferRollover) {
-    auto nack_packet = createNack(kRetransmissionsBufferSize - 1, VIDEO_PACKET);
+    uint ssrc = connection->getVideoSourceSSRC();
+    uint source_ssrc = connection->getVideoSinkSSRC();
+    auto nack_packet = erizo::Tools::createNack(ssrc, source_ssrc, kRetransmissionsBufferSize - 1, VIDEO_PACKET);
 
-    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(kRetransmissionsBufferSize))).Times(1);
-    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(kRetransmissionsBufferSize - 1))).Times(2);
-    pipeline->write(createDataPacket(kRetransmissionsBufferSize - 1, VIDEO_PACKET));
-    pipeline->write(createDataPacket(kRetransmissionsBufferSize, VIDEO_PACKET));
+    EXPECT_CALL(*writer.get(), write(_, _)).
+        With(Args<1>(HasSequenceNumber(kRetransmissionsBufferSize))).Times(1);
+    EXPECT_CALL(*writer.get(), write(_, _)).
+        With(Args<1>(HasSequenceNumber(kRetransmissionsBufferSize - 1))).Times(2);
+    pipeline->write(erizo::Tools::createDataPacket(kRetransmissionsBufferSize - 1, VIDEO_PACKET));
+    pipeline->write(erizo::Tools::createDataPacket(kRetransmissionsBufferSize, VIDEO_PACKET));
 
     EXPECT_CALL(*reader.get(), read(_, _)).Times(0);
     pipeline->read(nack_packet);
 }
 
 TEST_F(RtpRetransmissionHandlerTest, shouldRetransmitPackets_whenReceivingNackWithMultipleSeqNums) {
-    auto nack_packet = createNack(kArbitrarySeqNumber, VIDEO_PACKET, 1);
+    uint ssrc = connection->getVideoSourceSSRC();
+    uint source_ssrc = connection->getVideoSinkSSRC();
+    auto nack_packet = erizo::Tools::createNack(ssrc, source_ssrc, erizo::kArbitrarySeqNumber, VIDEO_PACKET, 1);
 
-    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(kArbitrarySeqNumber))).Times(2);
-    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(kArbitrarySeqNumber + 1))).Times(2);
-    pipeline->write(createDataPacket(kArbitrarySeqNumber, VIDEO_PACKET));
-    pipeline->write(createDataPacket(kArbitrarySeqNumber + 1, VIDEO_PACKET));
+    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(erizo::kArbitrarySeqNumber))).Times(2);
+    EXPECT_CALL(*writer.get(), write(_, _)).With(Args<1>(HasSequenceNumber(erizo::kArbitrarySeqNumber + 1))).Times(2);
+    pipeline->write(erizo::Tools::createDataPacket(erizo::kArbitrarySeqNumber, VIDEO_PACKET));
+    pipeline->write(erizo::Tools::createDataPacket(erizo::kArbitrarySeqNumber + 1, VIDEO_PACKET));
 
     EXPECT_CALL(*reader.get(), read(_, _)).Times(0);
     pipeline->read(nack_packet);
